Move some methods from rmg.react to data.kinetics.common
These methods are mostly used by kinetics database modules
This will alleviate some cyclic import issues

Also updates all related imports
