+++ rmgpy/data/kinetics/common.py
@@ -33,6 +33,7 @@ This module contains classes and functions that are used by multiple modules
 in this subpackage.
 """
 import itertools
+import logging
 
 from rmgpy.data.base import LogicNode
 from rmgpy.reaction import Reaction
@@ -224,3 +225,184 @@ def generate_molecule_combos(input_species):
         raise ValueError('Reaction generation can be done for 1 or 2 species, not {0}.'.format(len(input_species)))
 
     return combos
+
+
+def _labelListOfSpecies(speciesTuple):
+    """
+    given a list or tuple of species' objects, ensure all their atoms' id are
+    independent.
+
+    Modifies the speciesTuple in place, nothing returned.
+    """
+    # assert that all species' atomlabels are different
+    def independentIDs():
+        num_atoms = 0
+        IDs = []
+        for species in speciesTuple:
+            num_atoms += len(species.molecule[0].atoms)
+            IDs.extend([atom.id for atom in species.molecule[0].atoms ])
+        num_ID = len(set(IDs))
+        return num_ID == num_atoms
+    # if they are different, relabel and remake atomIDs
+    if not independentIDs():
+        logging.debug('identical atom ids found between species. regenerating')
+        for species in speciesTuple:
+            mol = species.molecule[0]
+            mol.assignAtomIDs()
+            # remake resonance isomers with new labeles
+            species.molecule = [mol]
+            species.generateResonanceIsomers(keepIsomorphic = True)
+
+
+def findDegeneracies(rxnList, sameReactants=None):
+    """
+    given a list of Reaction object with Molecule objects, this method
+    removes degenerate reactions and increments the degeneracy of the
+    reaction object. For multiple transition states, this method adds
+    them as separate duplicate reactions. This method modifies
+    rxnList in place and does not return anything.
+
+    This algorithm used to exist in family.__generateReactions, but was moved
+    here because it didn't have any family dependence.
+    """
+
+    # We want to sort all the reactions into sublists composed of isomorphic reactions
+    # with degenerate transition states
+    rxnSorted = []
+    for rxn0 in rxnList:
+        # find resonance structures for rxn0
+        convertToSpeciesObjects(rxn0)
+        if len(rxnSorted) == 0:
+            # This is the first reaction, so create a new sublist
+            rxnSorted.append([rxn0])
+        else:
+            # Loop through each sublist, which represents a unique reaction
+            for rxnList1 in rxnSorted:
+                # Try to determine if the current rxn0 is identical or isomorphic to any reactions in the sublist
+                isomorphic = False
+                identical = False
+                sameTemplate = False
+                for rxn in rxnList1:
+                    isomorphic = rxn0.isIsomorphic(rxn, checkIdentical=False, checkTemplateRxnProducts=True)
+                    if not isomorphic:
+                        identical = False
+                    else:
+                        identical = rxn0.isIsomorphic(rxn, checkIdentical=True, checkTemplateRxnProducts=True)
+                    sameTemplate = frozenset(rxn.template) == frozenset(rxn0.template)
+                    if not isomorphic:
+                        # a different product was found, go to next list
+                        break
+                    elif not sameTemplate:
+                        # a different transition state was found, mark as duplicate and
+                        # go to the next sublist
+                        rxn.duplicate = True
+                        rxn0.duplicate = True
+                        break
+                    elif identical:
+                        # An exact copy of rxn0 is already in our list, so we can move on to the next rxn
+                        break
+                    else: # sameTemplate and isomorphic but not identical
+                        # This is the right sublist for rxn0, but continue to see if there is an identical rxn
+                        continue
+                else:
+                    # We did not break, so this is the right sublist, but there is no identical reaction
+                    # This means that we should add rxn0 to the sublist as a degenerate rxn
+                    rxnList1.append(rxn0)
+                if isomorphic and sameTemplate:
+                    # We already found the right sublist, so we can move on to the next rxn
+                    break
+            else:
+                # We did not break, which means that there was no isomorphic sublist, so create a new one
+                rxnSorted.append([rxn0])
+
+    rxnList = []
+    for rxnList1 in rxnSorted:
+        # Collapse our sorted reaction list by taking one reaction from each sublist
+        rxn = rxnList1[0]
+        # The degeneracy of each reaction is the number of reactions that were in the sublist
+        rxn.degeneracy = sum([reaction0.degeneracy for reaction0 in rxnList1])
+        rxnList.append(rxn)
+
+    for rxn in rxnList:
+        if rxn.isForward:
+            reduceSameReactantDegeneracy(rxn, sameReactants)
+        else:
+            # fix the degeneracy of (not ownReverse) reactions found in the backwards direction
+            correctDegeneracyOfReverseReaction(rxn)
+
+    return rxnList
+
+
+def convertToSpeciesObjects(reaction):
+    """
+    modifies a reaction holding Molecule objects to a reaction holding
+    Species objects, with generated resonance isomers.
+    """
+    # if already species' objects, return none
+    if isinstance(reaction.reactants[0],Species):
+        return None
+    # obtain species with all resonance isomers
+    for i, mol in enumerate(reaction.reactants):
+        spec = Species(molecule = [mol])
+        if not reaction.isForward:
+            spec.generateResonanceIsomers(keepIsomorphic=True)
+        reaction.reactants[i] = spec
+    for i, mol in enumerate(reaction.products):
+        spec = Species(molecule = [mol])
+        if reaction.isForward:
+            spec.generateResonanceIsomers(keepIsomorphic=True)
+        reaction.products[i] = spec
+
+    # convert reaction.pairs object to species
+    newPairs=[]
+    for reactant, product in reaction.pairs:
+        newPair = []
+        for reactant0 in reaction.reactants:
+            if reactant0.isIsomorphic(reactant):
+                newPair.append(reactant0)
+                break
+        for product0 in reaction.products:
+            if product0.isIsomorphic(product):
+                newPair.append(product0)
+                break
+        newPairs.append(newPair)
+    reaction.pairs = newPairs
+
+    try:
+        convertToSpeciesObjects(reaction.reverse)
+    except AttributeError:
+        pass
+
+
+def reduceSameReactantDegeneracy(reaction, sameReactants=None):
+    """
+    This method reduces the degeneracy of reactions with identical reactants,
+    since translational component of the transition states are already taken
+    into account (so swapping the same reactant is not valid)
+
+    This comes from work by Bishop and Laidler in 1965
+    """
+    if len(reaction.reactants) == 2 and (
+                (reaction.isForward and sameReactants) or
+                reaction.reactants[0].isIsomorphic(reaction.reactants[1])
+            ):
+        reaction.degeneracy *= 0.5
+        logging.debug('Degeneracy of reaction {} was decreased by 50% to {} since the reactants are identical'.format(reaction, reaction.degeneracy))
+
+
+def correctDegeneracyOfReverseReaction(reaction):
+    """
+    This method corrects the degeneracy of reactions found when the backwards
+    template is used. Given the following parameters:
+
+        reaction - list of reactions with their degeneracies already counted
+
+    This method modifies reaction in place and returns nothing
+
+    This does not adjust for identical reactants, you need to use `reduceSameReactantDegeneracy`
+    to adjust for that.
+    """
+    from rmgpy.data.rmg import getDB
+    family = getDB('kinetics').families[reaction.family]
+    if not family.ownReverse:
+        reaction.degeneracy = family.calculateDegeneracy(reaction)
